Create textbox document filter for dollars
Created a subclass of DocumentFilter to create a filter to apply to textboxes meant to take in only dollar amounts. This filter will make sure only a maximum of two decimal places can be entered. The textbox will also only allow a maximum of 18 characters (For example, 15 digits, 1 decimal point, and 2 decimal digits).