Remove check for empty database
We had placed a check to see if the database does not exist before calling the method to create tables. This was redundant as the SQL that handles the creation of tables is set up to check if the table already exists or not. Therefore, there is no issue always calling the create tables method. Especially since as of now this application only has 2 tables.