Use method to initialize panels
Instead of initializing the JPanels where the attributes are declared, a method 'initPanels' is now responsible for initialization. This is beneficial as this now gurantees that a conneciton exists before the panels are initialized.

Also notice that a DatabaseConnection object is passed into the constructor of all of these panels.